+++ internal/compose/hover_test.go
@@ -505,6 +505,28 @@ services:
 				},
 			},
 		},
+		{
+			name: "service hover with whitespace around it",
+			content: `
+services:
+
+  backend:
+    image: hello
+
+  frontend:
+    depends_on:
+      - backend`,
+			line:      8,
+			character: 12,
+			result: &protocol.Hover{
+				Contents: protocol.MarkupContent{
+					Kind: protocol.MarkupKindMarkdown,
+					Value: "```YAML\n" + `backend:
+  image: hello` +
+						"\n```",
+				},
+			},
+		},
 	}
 
 	composeFile := fmt.Sprintf("file:///%v", strings.TrimPrefix(filepath.ToSlash(filepath.Join(os.TempDir(), "compose.yaml")), "/"))